Basic sprite strategy and mouse point to 3D
Add support for the basic sprite strategy in 3D and added a function to obtain a point in 3D where the mouse was clicked. Currently this point is only mapped to the xy plane and the camera has to be facing straight down the -z axis.

The next thing planned is to allow mouse projection onto a plane of choice while the camera is in any position.